+++ src/application/context/invariants/search/SearchInvariantsResponse.ts
@@ -1,0 +1,9 @@
+/**
+ * SearchInvariantsResponse - Response DTO for the search-invariants use case.
+ */
+
+import { InvariantView } from "../InvariantView.js";
+
+export interface SearchInvariantsResponse {
+  readonly invariants: InvariantView[];
+}